REF: use common class for backends
